Simple model with INST data manually run in INCA GUI
Reviewer asked for validation that the a GUI run gives the same results as the INCAWrapper. To accomudate this, we manually setup the simple model INST simulated dataset through the INCA GUI and ran estimation and parameter continuation. The model is save as simpale_model_inst_gui.mat and the resulting flux map is saved as simple_model_inst_gui_flux_map.mat.