Preparred for analysis with fixed initial values and rounded data.
We suspect that the small differences in the input due to different rounding causes small deviations in results. Therefore, in this commit the simulations scripts outputs data files were the data is rounded to 4 decimal points. The new data set was manully entered in to the inca model through the GUI. To be able to load the data through the INCAResults object before flux estimation is executed, we run and save a simulation of the model with the rounded data.